refactor(cli): router to a different package
