+++ API/main_test.go
@@ -1,13 +1,10 @@
 package main
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
 	"net/http"
-	"net/http/httptest"
 	"os"
-	"p3/app"
 	"p3/test/e2e"
 	u "p3/utils"
 	"reflect"
@@ -17,15 +14,6 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func makeRequest(method, url string, requestBody []byte) *httptest.ResponseRecorder {
-	router := Router(app.JwtAuthentication)
-	recorder := httptest.NewRecorder()
-	request, _ := http.NewRequest(method, url, bytes.NewBuffer(requestBody))
-	request.Header.Set("Authorization", "Bearer "+e2e.AdminToken)
-	router.ServeHTTP(recorder, request)
-	return recorder
-}
-
 func TestCreateLoginAccount(t *testing.T) {
 	// Test create new account
 	requestBody := []byte(`{
@@ -33,26 +21,26 @@ func TestCreateLoginAccount(t *testing.T) {
 		"password": "pass123secret",
 		"roles":{"*":"manager"}
 	}`)
-	recorder := makeRequest("POST", "/api/users", requestBody)
+	recorder := e2e.MakeRequest("POST", "/api/users", requestBody)
 	assert.Equal(t, http.StatusCreated, recorder.Code)
 
 	var response map[string]interface{}
 	json.Unmarshal(recorder.Body.Bytes(), &response)
 	_, exists := response["account"].(map[string]interface{})["token"].(string)
-	assert.Equal(t, true, exists)
+	assert.True(t, exists)
 
 	// Test recreate existing account
-	recorder = makeRequest("POST", "/api/users", requestBody)
+	recorder = e2e.MakeRequest("POST", "/api/users", requestBody)
 	println(recorder.Body.String())
 	assert.Equal(t, http.StatusBadRequest, recorder.Code)
 
 	// Test login
-	recorder = makeRequest("POST", "/api/login", requestBody)
+	recorder = e2e.MakeRequest("POST", "/api/login", requestBody)
 	assert.Equal(t, http.StatusOK, recorder.Code)
 
 	json.Unmarshal(recorder.Body.Bytes(), &response)
 	_, exists = response["account"].(map[string]interface{})["token"].(string)
-	assert.Equal(t, true, exists)
+	assert.True(t, exists)
 }
 
 func TestObjects(t *testing.T) {
@@ -74,33 +62,33 @@ func TestObjects(t *testing.T) {
 		data, _ = json.Marshal(obj)
 
 		// Create (POST)
-		recorder := makeRequest("POST", "/api/"+entStr+"s", data)
+		recorder := e2e.MakeRequest("POST", "/api/"+entStr+"s", data)
 
 		assert.Equal(t, http.StatusCreated, recorder.Code)
 		json.Unmarshal(recorder.Body.Bytes(), &response)
 		fmt.Println(response)
 		id, exists := response["data"].(map[string]interface{})["id"].(string)
-		assert.Equal(t, true, exists)
+		assert.True(t, exists)
 
 		// Verify create with GET
-		recorder = makeRequest("GET", "/api/"+entStr+"s/"+id, nil)
+		recorder = e2e.MakeRequest("GET", "/api/"+entStr+"s/"+id, nil)
 		assert.Equal(t, http.StatusOK, recorder.Code)
 		var responseGET map[string]interface{}
 		json.Unmarshal(recorder.Body.Bytes(), &responseGET)
 		delete(responseGET, "message")
 		delete(response, "message")
-		assert.Equal(t, true, reflect.DeepEqual(response, responseGET))
+		assert.True(t, reflect.DeepEqual(response, responseGET))
 
 		// Update with PUT
 		oldName := obj["name"].(string)
 		obj["name"] = entStr + "Test"
 		data, _ = json.Marshal(obj)
-		recorder = makeRequest("PUT", "/api/"+entStr+"s/"+id, data)
+		recorder = e2e.MakeRequest("PUT", "/api/"+entStr+"s/"+id, data)
 		assert.Equal(t, http.StatusOK, recorder.Code)
 
 		// Verify it
 		id = strings.Replace(id, oldName, obj["name"].(string), 1)
-		recorder = makeRequest("GET", "/api/"+entStr+"s/"+id, nil)
+		recorder = e2e.MakeRequest("GET", "/api/"+entStr+"s/"+id, nil)
 		assert.Equal(t, http.StatusOK, recorder.Code)
 		parentId = id
 	}
@@ -110,11 +98,11 @@ func TestObjects(t *testing.T) {
 	requestBody := []byte(`{
 		"name": "` + hierarchyName + `"
 	}`)
-	recorder := makeRequest("PATCH", "/api/sites/siteTest", requestBody)
+	recorder := e2e.MakeRequest("PATCH", "/api/sites/siteTest", requestBody)
 	assert.Equal(t, http.StatusOK, recorder.Code)
 
 	// Verify the whole tree
-	recorder = makeRequest("GET", "/api/sites/"+hierarchyName+"/all", nil)
+	recorder = e2e.MakeRequest("GET", "/api/sites/"+hierarchyName+"/all", nil)
 	assert.Equal(t, http.StatusOK, recorder.Code)
 	json.Unmarshal(recorder.Body.Bytes(), &response)
 	response = response["data"].(map[string]interface{})
@@ -130,9 +118,9 @@ func TestObjects(t *testing.T) {
 	}
 
 	// Delete everything
-	recorder = makeRequest("DELETE", "/api/sites/TESTPATCH", nil)
+	recorder = e2e.MakeRequest("DELETE", "/api/sites/TESTPATCH", nil)
 	assert.Equal(t, http.StatusNoContent, recorder.Code)
-	recorder = makeRequest("GET", "/api/hierarchy", nil)
+	recorder = e2e.MakeRequest("GET", "/api/hierarchy", nil)
 	assert.Equal(t, http.StatusOK, recorder.Code)
 	json.Unmarshal(recorder.Body.Bytes(), &response)
 	assert.Nil(t, response["data"].(map[string]interface{})["tree"].(map[string]interface{})["physical"].(map[string]interface{})["TESTPATCH"])
